requirements/dev.txt
# Dependencies needed only for development/testing.

# for the test suite and coverage metrics
# sha256: gpJO-16ng6ciNGgqDoBJ2E9erryqw8jAiTt-rpfyg4A
pytest==2.7.2
# sha256: MeUZjAbykOHoE9MN3QY6BAHZ46aYEpykQmd4FnOnEoU
pytest-cov==1.8.1
# sha256: 8UAnAQY5mPanBgl4mq6NwFcD860KNIgvYZllNlTFVUM
datadiff==1.1.6

# sha256: vrVp5tcHy0zYLCJsl7fh9BzWK65lbYnmhxB-74N2x_0
WebTest==2.0.18

# sha256: uDndLZwRfHAUMMFJlWkYpCOphjtIsJyQ4wpgE-fS9E8
mock==1.0.1
# sha256: pd97XgYOqpGdfrd_rqzwrMfhAHS7NMmgrCI44hhRRzs
responses==0.4.0

# sha256: PJgkO9orHrHMxj21phTwzqy_QBRMkrN8W6dVzsu8wN4
django-extensions==1.5.5

#for celery auto-reloading
# sha256: nJmKXXYGyoNQZc2rwBOubGbrnqdqAKHjvG4M_itPcfQ
pyinotify==0.9.6

# sha256: j8ToZaKfPG7K5AGn3OGCR_ir0EbSTjkOFq-Zdsyceqg
flake8==2.4.1

# Required by pytest
# sha256: B-IKuQpVC9PCGJHg2IfwkxtAmPFIrsleKbUYjxYbsHU
py==1.4.30

# Required by pytest-cov
# sha256: ShTGfVIP2p1CsNphNGOFeMquHTdLm7Ri2N4AWH26dkw
cov-core==1.15.0

# Required by cov-core
# sha256: 0a6hxKphuDZtakLdNlBiL7-cY07STq9_N5yLlw5e1E4
coverage==3.7.1

# Required by WebTest
# sha256: twJyH4ciyU-9L737Qhg8S1s9iHEYa5KoiY4t_khpqts
beautifulsoup4==4.4.0
# sha256: gmUn3J0zTtTtds2uZy_cu8z2FBhmV9txZ5q1jfhpRYo
waitress==0.8.9
# sha256: Evi5g5C-_EczbSwOW62cxIYJ2Ajquz-GddwQJ6Op6ds
WebOb==1.4.1

# Required by responses
# sha256: Fb7nUwAt_2hJh7jfjCNSiOuNRfgZGuBWJUgS39QsgdM
cookies==2.2.1

# Required by django-extensions, WebTest & responses
# sha256: QYqTw5en7asj5ViNvAZ6x0pyPts9VBvUk295R252Rdo
six==1.9.0

# Required by flake8
# sha256: vWwID7NyrrywzhnjXdrHRPKr9ae--iB9stEJfUjv5jo
mccabe==0.3.1
# sha256: Yuh_1UU1-5MrSk2Uho21IyV6EDHIvGvTWMIBVDPmRts
pep8==1.5.7
# sha256: Bd9YSinu6poqIRDdNi5T0E4MS7F1S01xI09lGRfzwvA
pyflakes==0.9.2
